+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Allow forcing a specific timestamp value via `SR_TIMESTAMP_OVERRIDE` env var. This helps with reproducibility and testing.
+
 ## v0.5.0 (2022-08-17)
 
 - Make labels not required by default