+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Allow forcing a specific timestamp value via `SR_TIMESTAMP_OVERRIDE` env var. This helps with reproducibility and testing.
+- Use CRLF line endings for text files on Windows
 
 ## v0.5.0 (2022-08-17)