+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Allow forcing a specific timestamp value via `SR_TIMESTAMP_OVERRIDE` env var. This helps with reproducibility and testing.
 - Use CRLF line endings for text files on Windows
 - Add json_schema_url property to label definitions
+- Validate label answers against the label's JSON schema
 
 ## v0.5.0 (2022-08-17)